backend/requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cors==1.0.1
dotenv==0.9.9
filelock==3.18.0
Flask==3.1.0
Flask-JWT-Extended==4.7.1
future==1.0.0
gevent==25.5.1
greenlet==3.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
PyJWT==2.10.1
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
requests-file==2.1.0
setuptools==80.9.0
tldextract==5.3.0
urllib3==2.4.0
Werkzeug==3.1.3
zope.event==5.0